refactor(diagnostics): give browser samples one validation and one projection owner
The browser sample contract was described three times: the renderer built the
payload, the route validated it inline, and RuntimeDiagnosticsService applied a
second set of allowlists on the way to storage. Unknown fields behaved
differently at each boundary, and a nested key rejected with a 422 while an
unknown top-level key was dropped in silence.

BrowserDiagnosticSampleRequest now owns the accepted shape, the payload byte
budget, and unknown fields at every depth. BrowserDiagnosticSampleFormatter owns
what happens on the way to storage: the URL reduced to a redacted path plus a
hash, animation detail capped to the configured limits, and empty sections
dropped. The renderer stays a separate producer, held to the request's field list
by a parity test.

Closes #165
